+++ test_more/general/div_eq.cpp
@@ -122,7 +122,7 @@ bool DivEqTestTwo(void)
 	double value = 1./8.;
 	v[0]         = 1.;
 	for(j = 1; j < p; j++)
-	{	jfac *= j;
+	{	jfac *= double(j);
 		w     = f.Forward(j, v);
 		ok &= NearEqual(w[0], value/jfac, eps99, eps99); // d^jz/du^j
 		v[0]  = 0.;
@@ -137,7 +137,7 @@ bool DivEqTestTwo(void)
 	value = 1./8.;
 	for(j = 0; j < p; j++)
 	{	ok &= NearEqual(r[j], value/jfac, eps99, eps99); // d^jz/du^j
-		jfac *= (j + 1);
+		jfac *= double(j + 1);
 		value = 0.;
 	}
